--- client/src/components/Header/Header.css
@@ -1,3 +1,0 @@
-nav {
-  margin-bottom: 3rem;
-}
